feat: 상품 리스트 조회 기능 추가

+++ module-common/src/main/java/com/example/common/response/PageResponse.java
@@ -9,10 +9,10 @@ public class PageResponse<T> {
 	private List<T> content;
 	private int page;
 	private int size;
-	private int totalElements;
+	private long totalElements;
 	private int totalPages;
 
-	private PageResponse(List<T> content, int page, int size, int totalElements) {
+	private PageResponse(List<T> content, int page, int size, long totalElements) {
 		this.content = content;
 		this.page = page;
 		this.size = size;
@@ -21,7 +21,7 @@ public class PageResponse<T> {
 		this.totalPages = (int) ((totalElements + size - 1) / size);
 	}
 
-	public static <T> PageResponse<T> of(List<T> content, int page, int size, int totalElements) {
+	public static <T> PageResponse<T> of(List<T> content, int page, int size, long totalElements) {
 		return new PageResponse<>(content, page, size, totalElements);
 	}
 }
